Move lint_map_unwrap_or to its own file
